pyproject.toml
[tool.poetry]
name = "beard_drf"
version = "0.1.0"
description = ""
authors = ["Isa Suleymanov <issashkalikimike@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
